refactor(lm): decompose add-time photo import phase
